Refactoring + Implementasi awal GCM
