+++ app/src/main/java/picklenostra/user_app/Helper/LruBitmapCache.java
@@ -1,4 +1,4 @@
-package picklenostra.user_app;
+package picklenostra.user_app.Helper;
 
 /**
  * Created by Edwin on 4/21/2016.
