+++ app/src/main/java/picklenostra/user_app/SearchWithMapActivity.java
@@ -5,11 +5,11 @@ import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.graphics.Color;
 import android.location.Location;
+import android.media.MediaPlayer;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.ActivityCompat;
 import android.support.v7.app.ActionBarActivity;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.MenuItem;
@@ -45,7 +45,10 @@ import java.math.RoundingMode;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.List;
+
+import picklenostra.user_app.Adapter.SearchWithMapAdapter;
+import picklenostra.user_app.Helper.VolleyController;
+import picklenostra.user_app.Model.SearchWithMapModel;
 
 public class SearchWithMapActivity extends ActionBarActivity implements
         LocationListener, GoogleApiClient.ConnectionCallbacks, GoogleApiClient.OnConnectionFailedListener {
@@ -54,6 +57,7 @@ public class SearchWithMapActivity extends ActionBarActivity implements
     private GoogleApiClient googleApiClient;
     private Location location;
     private LocationRequest locationRequest;
+    private MediaPlayer mediaPlayer;
     private double curLat, curLong;
 
     private ListView listView;
@@ -165,13 +169,6 @@ public class SearchWithMapActivity extends ActionBarActivity implements
         locationRequest.setInterval(INTERVAL);
 
         if (ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-            // TODO: Consider calling
-            //    ActivityCompat#requestPermissions
-            // here to request the missing permissions, and then overriding
-            //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-            //                                          int[] grantResults)
-            // to handle the case where the user grants the permission. See the documentation
-            // for ActivityCompat#requestPermissions for more details.
             return;
         }
         LocationServices.FusedLocationApi.requestLocationUpdates(googleApiClient, locationRequest, this);
